Cleanup the external interface and split files
Switch more-example to async server

client: capture message by value

Update readme for asio
