+++ include/varlink/detail/config.hpp
@@ -1,0 +1,8 @@
+#ifndef LIBVARLINK_CONFIG_HPP
+#define LIBVARLINK_CONFIG_HPP
+#include <asio.hpp>
+
+namespace varlink {
+namespace net = ::asio;
+}
+#endif // LIBVARLINK_CONFIG_HPP
