+++ tests/cert_client.cpp
@@ -7,7 +7,8 @@ int main(int argc, char* argv[]) {
         std::cerr << "Error: varlink socket path required\n";
         return 1;
     }
-    auto client = varlink::varlink_client(argv[1]);
+    varlink::net::io_context ctx{};
+    auto client = varlink::varlink_client(ctx, argv[1]);
 
     try {
         auto resp = client.call("org.varlink.certification.Start", {})();
