+++ tests/self_cert.sh
