+++ tests/self_unix.hpp
@@ -33,10 +33,12 @@ class UnixEnvironment : public BaseEnvironment {
     UnixEnvironment() : BaseEnvironment()
     {
         std::filesystem::remove(get_endpoint().path());
-        server = std::make_unique<test_server>(varlink_uri, description);
 #ifdef VARLINK_TEST_ASYNC
+        server = std::make_unique<test_server>(ctx, varlink_uri, description);
         timer = std::make_unique<net::steady_timer>(server->get_executor());
-        worker = std::thread([&]() { server->run(); });
+        worker = std::thread([&]() { ctx.run(); });
+#else
+        server = std::make_unique<test_server>(varlink_uri, description);
 #endif
     }
 };
